msm: mdss: change the print level to debug for dynamic fps write
During stability testing, the device can enter and exit the idle
screen use case very frequently. When this happens the fps switch
between the minimum and maximum configured values happens many
times causing the kernel logs to be flooded with messages related
to this. Hence, silence this print statement by changing it to
pr_debug.

Change-Id: If54871c1d27fcc50d901712c5fd5f279a8bd3efc
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>